backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pypdf
numpy
faiss-cpu
requests
python-multipart